refactor forms into their own directory + MainNavigation renamed to Headers
