+++ src/components/layout/Layout.js
@@ -1,11 +1,11 @@
-import MainNavigation from "./MainNavigation";
+import Headers from "./Headers";
 import Footers from "./Footers";
 import classes from './Layout.module.css';
 
 function Layout(props) {
     return (
         <div>
-            <MainNavigation />
+            <Headers />
             <main className={classes.main}>{props.children}</main>
             <Footers />
         </div>
